requirements.txt
Flask==2.0.3
google-cloud-datastore==1.15.3
python-jose
flask-cors
six
python-dotenv
requests
authlib
google-api-python-client
google-auth 
google-auth-oauthlib 
google-auth-httplib2